+++ src/lib/table-tooltip/index.ts
@@ -1,0 +1,4 @@
+export * from './table-tooltip.types';
+export * from './table-tooltip.token';
+export * from './table-tooltip.provider';
+export * from './table-tooltip.directive';